+++ bitcoind-rpc/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for bitcoind-rpc
 
+## 0.2.0.0 -- ?
+
+* Define our own `BlockHeader` type instead of using the one from `haskoin-core`
+* Refactor code to deal with new `aeson` instance definitions in `haskoin-core 0.13.5`
+
 ## 0.1.0.0 -- 2020-03-07
 
 Initial release